Update OBJECTS_AS_THINGS.md
I might be wrong, but I think .clear() is not a javascript array method? It is currently used in the "puke" method in the Rufus the dog object (twice- once using 'this' and once using the object name). I changed it to use .length = 0 as a possible alternative to clearing the contentsOfStomach array.